+++ Controls-demo/Slider/Base/Template.wml
@@ -67,13 +67,5 @@
             </Router.router:Reference>
             <Controls-demo.Slider.Base.Intervals.Index/>
         </div>
-        <div class="controlsDemo__cell">
-            <Router.router:Reference attr:class="controls-text-label"
-                                     state="/Controls-demo/app/:app"
-                                     app="{{ 'Controls-demo/Slider/Base/Intervals/Index' }}">
-                <a href="{{content.href}}">Получает значение из инпута</a>
-            </Router.router:Reference>
-            <Controls-demo.Slider.Base.withInput.Index/>
-        </div>
     </div>
 </div>
