--- Controls-demo/Slider/Base/withInput/Template.wml
@@ -1,13 +1,0 @@
-<div class="controlsDemo__wrapper controlsDemo_fixedWidth500">
-    <div class="controlsDemo__cell">
-        <div class="controls-text-label">with input</div>
-        <Controls.slider:Base maxValue="{{100}}"
-                              minValue="{{0}}"
-                              bind:value="_value"/>
-    </div>
-    <div class="controlsDemo__cell">
-        <Controls.input:Number bind:value="_value"
-                               attr:style="width:100px"
-                               on:valueChanged="_changedHandler()"/>
-    </div>
-</div>
